+++ src/render/primitives/ConvexOperation.cpp
@@ -62,8 +62,10 @@ PrimitivePacketHit4
 ConvexOperation::intersectPacketHits(const Ray4& rays, const PrimitivePacketState4& states) const {
   PrimitivePacketHit4 result;
   for (std::size_t lane = 0; lane != Ray4::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersectConvexRay(rays.rayd(lane), hitPoints, state);
     const HitPoint hitPoint = hitPoints.minWithPositiveDistance();
@@ -78,8 +80,10 @@ PrimitivePacketHit8
 ConvexOperation::intersectPacketHits(const Ray8& rays, const PrimitivePacketState8& states) const {
   PrimitivePacketHit8 result;
   for (std::size_t lane = 0; lane != Ray8::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersectConvexRay(rays.rayd(lane), hitPoints, state);
     const HitPoint hitPoint = hitPoints.minWithPositiveDistance();
@@ -95,8 +99,10 @@ ConvexOperation::intersectPacketIntervals(const Ray4& rays,
                                           const PrimitivePacketState4& states) const {
   PrimitivePacketInterval4 result;
   for (std::size_t lane = 0; lane != Ray4::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersectConvexRay(rays.rayd(lane), hitPoints, state);
     if (primitive || !hitPoints.empty()) {
@@ -111,8 +117,10 @@ ConvexOperation::intersectPacketIntervals(const Ray8& rays,
                                           const PrimitivePacketState8& states) const {
   PrimitivePacketInterval8 result;
   for (std::size_t lane = 0; lane != Ray8::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersectConvexRay(rays.rayd(lane), hitPoints, state);
     if (primitive || !hitPoints.empty()) {