+++ src/render/primitives/Disk.cpp
@@ -39,8 +39,10 @@ template<typename Packet, typename StateArray, typename Result>
 Result Disk::intersectPacketHitsFor(const Packet& rays, const StateArray& states) const {
   Result result;
   for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     const Rayd ray = rays.rayd(lane);
     const double t = (m_center - ray.origin()) * m_normal / (ray.direction() * m_normal);
     const Vector4d hitPoint = ray.at(t);