+++ src/render/primitives/Instance.cpp
@@ -185,8 +185,10 @@ Instance::intersectMovingRay4PacketHits(const Ray4& rays,
                                         const PrimitivePacketState4& states) const {
   PrimitivePacketHit4 result;
   for (std::size_t lane = 0; lane != Ray4::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersect(rays.rayd(lane), hitPoints, state);
     const HitPoint hitPoint = hitPoints.minWithPositiveDistance();
@@ -202,8 +204,10 @@ Instance::intersectMovingRay8PacketHits(const Ray8& rays,
                                         const PrimitivePacketState8& states) const {
   PrimitivePacketHit8 result;
   for (std::size_t lane = 0; lane != Ray8::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval hitPoints;
     const Primitive* primitive = intersect(rays.rayd(lane), hitPoints, state);
     const HitPoint hitPoint = hitPoints.minWithPositiveDistance();
@@ -219,8 +223,10 @@ Instance::intersectMovingRay4PacketIntervals(const Ray4& rays,
                                              const PrimitivePacketState4& states) const {
   PrimitivePacketInterval4 result;
   for (std::size_t lane = 0; lane != Ray4::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval interval;
     const Primitive* primitive = intersect(rays.rayd(lane), interval, state);
     if (primitive || !interval.empty()) {
@@ -239,8 +245,10 @@ Instance::intersectMovingRay8PacketIntervals(const Ray8& rays,
                                              const PrimitivePacketState8& states) const {
   PrimitivePacketInterval8 result;
   for (std::size_t lane = 0; lane != Ray8::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     HitPointInterval interval;
     const Primitive* primitive = intersect(rays.rayd(lane), interval, state);
     if (primitive || !interval.empty()) {