+++ src/render/primitives/Plane.cpp
@@ -88,8 +88,10 @@ template<typename Packet, typename StateArray, typename Result>
 Result Plane::intersectPacketHitsFor(const Packet& rays, const StateArray& states) const {
   Result result;
   for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     const Rayd ray = rays.rayd(lane);
     const double t = calculateIntersectionDistance(ray);
     if (t > 0.0) {