+++ src/render/primitives/Rectangle.cpp
@@ -53,8 +53,10 @@ template<typename Packet, typename StateArray, typename Result>
 Result Rectangle::intersectPacketHitsFor(const Packet& rays, const StateArray& states) const {
   Result result;
   for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     const Rayd ray = rays.rayd(lane);
     const double t = (m_corner - ray.origin()) * m_normal / (ray.direction() * m_normal);
     if (std::isinf(t)) {