+++ src/render/primitives/Torus.cpp
@@ -41,8 +41,10 @@ template<typename Packet, typename StateArray, typename Result>
 Result Torus::intersectPacketHitsFor(const Packet& rays, const StateArray& states) const {
   Result result;
   for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     const Rayd ray = rays.rayd(lane);
 
     if (!boundingBoxIntersects(ray)) {
@@ -81,8 +83,10 @@ template<typename Packet, typename StateArray, typename Result>
 Result Torus::intersectPacketIntervalsFor(const Packet& rays, const StateArray& states) const {
   Result result;
   for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
-    State fallbackState;
-    State& state = states[lane] ? *states[lane] : fallbackState;
+    if (!states[lane]) {
+      continue;
+    }
+    State& state = *states[lane];
     const Rayd ray = rays.rayd(lane);
 
     if (!boundingBoxIntersects(ray)) {